go.mod: increase the minimun golang version
Our CI system is installing the golang 1.15 for testing our code, while the
go.mod is pointing to 1.13, which isn't enough anymore for us. A user
running go1.13 won't be able to build the source code due to the following
error:

internal/logger/logger.go:50:61: undefined: log.Lmsgprefix

since this flag was added only in 1.14.
Since go1.15 is said as required in the docs and CI file, lets update the
go.mod version.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>