ci_run: deprecating --project option
The `lab ci create` and `lab ci trigger` are the only commands accepting the
--project/-p option, meaning that the user being in one git repository
directory can run/trigger a pipeline in any other project. This behavior is
quite inconsistent to the rest of the tool, where the project name and ID
are always inferred from the current directory/branch/remote where the user
stands at.

This patch deprecates the option to keep the consistency among other
options. It'll be left in hidden from the help/usage message to avoid
breaking backwards-compatibility, since this option is around since 2018.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>