ci_run: allow user to create merge requests pipeline
So far, lab only supported project-level pipelines, but left merge requests
pipelines out. This patch allows the user to run:

```
$ lab ci --merge-request <remote> <mr_IID>
```

Following the same convention as used by other merge request commands.
However, it's not possible to run it using the `--project` option, since,
on behalf of the CLI args, all other commands check for the the project the
user is checkout out at the moment.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>