ci_run: remove getCIRunOptions() function
The getCIRunOptions() function acts, partially, in the same way as the
parseArgsRemoteAndBranch() function when trying to get what's the correct
branch and remote name given by the user or guessed from the current working
branch, with that, the getCIRunOptions had left only the part where the
project ID is retrieved.

The project ID not always is necessary to be retrieved, since the go-gitlab
library interface and GitLab API accepts both an integer (the actual ID) or
a string (the name) of a project when performing any actions. Because of
that, the default value (from the working branch) can be the remote name
returned by parseArgsRemoteAndBranch(), while, only when the user uses the
--project option (now deprecated) we need to make sure the project exists,
then the ID is used instead (seems preferrable to avoid convertion within
Gitlab's API(?)).

Considering the getCIRunOptions() can be simplyfied, and also the fact we're
preparing the ground for allowing merge-request CI creation, which will add
some conditionals, it's simpler to just remove the function completely and
move the code to the commands themselves.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>