cmd: change --file behavior and add --no-edit
The --file (-F) option always considered the file input as the final version
of should be used as an issue comment or an MR description. However, on
mr_create command there is the flag --file-edit (-f), which allowed users to
edit the file before lab submitting it as the message.

The idea of this patch is to make the default behavior of --file to open the
file and only submit it without any edition when --no-edit is specified,
dropping the need for the --file-edit on mr_create. This change allow a
consistent experience among all subcommands with --no-edit available (which
should be for all those with --file available), thus improving user
experience.

Also, the --no-edit flag can be seen as a generic option of "no edition" of
any sort. Today, only --file seems a good fit, but later, --no-edit might be
useful to prevent any sort of "additional" edition that lab might offer to
the user.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>